Create aio dir for async channels
